create home layout
