+++ src/App.jsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import {useState} from "react";
 import "./App.css";
 import Footer from "./components/Footer/Footer";
