+++ src/pages/Home/Home.jsx
@@ -1,0 +1,11 @@
+import Button from "../../components/Button/Button";
+import React from "react";
+export default function Landing() {
+  return (
+    <>
+      <main>
+        <Button msg="View Avalaible" link="/reserva" />
+      </main>
+    </>
+  );
+}
