--- src/pages/landingPage/Landing.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Landing() {
-  return <div>Landing</div>;
-}
